+++ checkpoint/CHANGELOG.md
@@ -6,9 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Fixed
 - Return empty dict (instead of None) as metadata for *empty* Mapping typed
  unregistered types.
+
+ ### Changed
+- Removed `write_chunk_shape` and `read_chunk_shape` from `SaveArgs`.
 
 
 ## [0.6.3] - 2024-09-10